Changed state loop transition behaviour + fixed bug in example
A state returning to itself through a transition (either directly or
through a group/parent state) will now no longer have is entry or exit
actions triggered.

The various actions in the example have been fixed so that the example
follows the illustration. It is also changed so that it also
demonstrates entryState use, and how a state looping to itself doesn't
trigger entryActions/exitActions.